+++ docs/changelog.md
@@ -17,6 +17,7 @@ nav_order: 2
 
 ### Fixed
 
+- Fix Default and GitHub themes "auto" mode that wasn't picking the right colours when OS was set to dark.
 - Docusaurus theme colours and styling updated to be more faithful to the original.